+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 All notable changes to this project will be documented in this file.
 
+## [Unreleased]
+- configuration layer `AnyCache.configure`: an ability to choose and configure a necessary cache client
+  without any explicit client object instantiation (client object will be instantiated implicitly);
+
 ## [0.1.0] - 2018-08-26
 - Release :)