+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 # Changelog
 All notable changes to this project will be documented in this file.
 
-## [Unreleased]
+## [0.2.0] - 2018-09-03
 - fetching operation `AnyCache#fetch(key, force:, expires_in:, &block)`
   - fetches data from the cache using the given key;
   - if a block has been passed and data with the given key does not exist -